@@ -1,5 +1,3 @@
 module github.com/slomek/mappy
 
-go 1.12
-
-require github.com/pkg/errors v0.8.1
+go 1.16
