@@ -1,20 +1,19 @@
 package mappy
 
 import (
-	stderr "errors"
+	"errors"
+	"fmt"
 	"reflect"
-
-	"github.com/pkg/errors"
 )
 
 // ErrMapMarshal is returned when it is not possible to unmarshal map into a struct.
-var ErrMapUnmarshal = stderr.New("failed to unmarshal map into struct")
+var ErrMapUnmarshal = errors.New("failed to unmarshal map into struct")
 
 // Unmarshal transforms a string->string map into a custom struct.
 func Unmarshal(m map[string]string, data interface{}) (err error) {
 	defer func() {
 		if r := recover(); r != nil {
-			err = errors.Wrapf(ErrMapUnmarshal, "%v", r)
+			err = fmt.Errorf("%v: %w", r, ErrMapUnmarshal)
 		}
 	}()
 
